moved decorator in util/misc.py
